# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: QuantaWatch Admin API
  description: |
    The QuantaWatch Admin API provides observability and management endpoints
    for the QuantaWatch post-quantum security gateway.  It runs on a separate
    port (default 9091) from the proxy gateway (default 9090).

    All responses use JSON.  Authentication is not required by default but can
    be enabled in the gateway configuration.
  version: "0.1.0"
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: QuantaWatch
    url: https://github.com/dyber-inc/quantawatch
servers:
  - url: http://localhost:9091
    description: Local development (admin API)
paths:
...
  /api/sessions/{id}:
    get:
      operationId: getSession
      summary: Get session details
      description: Retrieve full details for a single session by its ID.
      tags:
        - Sessions
      parameters:
        - name: id
          in: path
          required: true
          description: Unique session identifier.
          schema:
            type: string
      responses:
        "200":
          description: Session details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "404":
          description: Session not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/audit:
    get:
      operationId: listAuditEntries
      summary: List audit entries
      description: |
        Returns entries from the tamper-evident, cryptographically signed
        audit log.  Entries are ordered by sequence number (ascending).
      tags:
        - Audit
      parameters:
        - name: limit
          in: query
          required: false
          description: Maximum number of entries to return.
          schema:
            type: integer
            minimum: 1
            default: 100
      responses:
        "200":
          description: A list of audit entries.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEntry"
...
components:
  schemas:
    SessionInfo:
      type: object
      description: An active or historical gateway session.
      required:
        - session_id
        - agent_name
        - created_at
        - expires_at
        - request_count
        - total_tokens
        - public_key_fingerprint
      properties:
        session_id:
          type: string
          description: Unique identifier for the session.
          examples:
            - "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        agent_name:
          type: string
          description: Name of the AI agent that owns the session.
          examples:
            - "support-bot"
        created_at:
          type: string
          format: date-time
          description: When the session was created (ISO 8601).
          examples:
            - "2025-06-01T10:00:00Z"
        expires_at:
          type: string
          format: date-time
          description: When the session expires (ISO 8601).
          examples:
            - "2025-06-01T11:00:00Z"
        request_count:
          type: integer
          description: Number of requests made in this session.
          examples:
            - 15
        total_tokens:
          type: integer
          description: Total tokens consumed in this session.
          examples:
            - 8200
        public_key_fingerprint:
          type: string
          description: Fingerprint of the ML-DSA public key used to create the session.
          examples:
            - "ML-DSA-65:ab12cd34ef56..."
    AuditEntry:
      type: object
      description: A single entry in the tamper-evident audit log.
      required:
        - sequence
        - timestamp
        - event_type
        - session_id
        - details
        - hash
        - signature
      properties:
        sequence:
          type: integer
          description: Monotonically increasing sequence number.
          examples:
            - 42
        timestamp:
          type: string
          format: date-time
          description: When the event occurred (ISO 8601).
          examples:
            - "2025-06-01T10:05:00Z"
        event_type:
          type: string
          description: |
            Type of event.  Common values: `session_created`,
            `request_proxied`, `threat_detected`, `session_expired`.
          examples:
            - "request_proxied"
        session_id:
          type: string
          description: Session this event belongs to.
          examples:
            - "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        details:
          type: object
          additionalProperties: true
          description: Arbitrary event-specific metadata.
          examples:
            - { "method": "POST", "path": "/v1/messages" }
        hash:
          type: string
          description: SHA3-256 hash chaining this entry to the previous one.
          examples:
            - "3a7f1b2c..."
        signature:
          type: string
          description: ML-DSA signature over the entry contents.
          examples:
            - "ML-DSA:a1b2c3..."
...
    ErrorResponse:
      type: object
      description: Standard error response.
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code.
          examples:
            - "not_found"
        message:
          type: string
          description: Human-readable error message.
          examples:
            - "Session not found"